refactor: split Update Subscription modal into per-tab forms
Extract Product, Discount, Trial, and Billing Period into dedicated
components so the modal is a thin tab shell. Product tab keeps the
existing Select UI for a follow-up picker change.
